+++ src/devices/smlight.ts
@@ -1,5 +1,6 @@
 import {Definition} from '../lib/types';
 import fz from '../converters/fromZigbee';
+import {forcePowerSource} from '../lib/modernExtend';
 
 const definitions: Definition[] = [
     {
@@ -10,6 +11,7 @@ const definitions: Definition[] = [
         fromZigbee: [fz.linkquality_from_basic],
         toZigbee: [],
         exposes: [],
+        extend: [forcePowerSource({powerSource: 'Mains (single phase)'})],
     },
 ];